+++ app/Modules/Products/DTOs/UpdateProductDTO.php
@@ -16,4 +16,9 @@ class UpdateProductDTO extends BaseDTO
     ) {
     }
 
+    public function toArray(): array
+    {
+        return array_filter(parent::toArray(), fn($value) => $value !== null);
+    }
+
 }