+++ config/app.php
@@ -36,6 +36,7 @@ return [
         App\Modules\Products\Providers\ProductsServiceProvider::class,
         App\Modules\Cart\Providers\CartServiceProvider::class,
         App\Modules\Orders\Providers\OrdersServiceProvider::class,
+        App\Modules\Coupons\Providers\CouponsServiceProvider::class,
     ],
 
     'aliases' => [